+++ tests/test_coverage.py
@@ -9,7 +9,7 @@ import difflib
 
 import pytest
 
-from finaletoolkit.frag.coverage import *
+from finaletoolkit.frag.coverage import coverage, single_coverage
 
 class TestSingleCoverage:
     def test_coverage(self, request):
@@ -45,8 +45,13 @@ class TestCoverage:
     def test_coverage_normalize(self, request):
         input_file = request.path.parent / 'data' / '12.3444.b37.frag.gz'
         intervals = request.path.parent / 'data' / 'intervals.bed'
-        results = coverage(input_file ,intervals, "-", scale_factor=1.,
-                           normalize=True, verbose=True)
+        results = coverage(
+            input_file,
+            intervals,
+            "-",
+            scale_factor=1.,
+            normalize=True,
+            verbose=True)
         for i in range(2):
             chrom, start, stop, name, cov = results[i]
             if i == 0:  
@@ -65,7 +70,7 @@ class TestCoverage:
     def test_coverage_no_normalize(self, request):
         input_file = request.path.parent / 'data' / '12.3444.b37.frag.gz'
         intervals = request.path.parent / 'data' / 'intervals.bed'
-        results = coverage(input_file,intervals,"-", normalize=False,
+        results = coverage(input_file, intervals, "-", normalize=False,
                            intersect_policy='midpoint',
                            scale_factor=1.)
         for i in range(2):